\name{soql-internal}
\alias{soql-internal}
\alias{as.character.soql}
\alias{print.soql}
\title{
Internal Functions
}
\description{
These functions are for use internally only.
}
\usage{
\method{as.character}{soql}(x, ...)
\method{print}{soql}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  \code{soql} object to print or convert to URL.
}
}
\examples{
as.character.soql(soql())
print.soql(soql())
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
